requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cleanup==9.0.0
djangorestframework==3.16.1
imageio==2.37.0
joblib==1.5.2
lazy_loader==0.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sqlparse==0.5.3
threadpoolctl==3.6.0
tifffile==2025.9.9
psycopg2-binary==2.9.10
dotenv~=0.9.9
python-dotenv~=1.1.1
dj-config-url~=0.1.1
gunicorn==23.0.0
django-storages[s3]==1.14.4